AppCrashTracker used to catch unhandled exceptions in Android
It is a kind of toolkit to track the exception arising in the
application and it will generate a json and can upload in your server
using your own post url.

No need to worry and think to add more line of code. Simple is better.
So a single invoke is enough to make better app.